+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/ProjectSettings.kt
@@ -1,11 +1,13 @@
 package com.github.blarc.ai.commits.intellij.plugin.settings
 
 import com.github.blarc.ai.commits.intellij.plugin.AICommitsUtils
+import com.github.blarc.ai.commits.intellij.plugin.settings.clients.LLMClientConfiguration
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil
+import com.intellij.util.xmlb.annotations.Attribute
 
 @State(
         name = ProjectSettings.SERVICE_NAME,
@@ -20,6 +22,11 @@ class ProjectSettings : PersistentStateComponent<ProjectSettings?> {
 
     var projectExclusions: Set<String> = setOf()
 
+    @Attribute
+    var activeLlmClientId: String? = null
+    @Attribute
+    var isProjectSpecificLLMClient: Boolean = false
+
     override fun getState() = this
 
     override fun loadState(state: ProjectSettings) {
@@ -30,5 +37,8 @@ class ProjectSettings : PersistentStateComponent<ProjectSettings?> {
         return AICommitsUtils.matchesGlobs(path, projectExclusions)
     }
 
+    fun getActiveLLMClientConfiguration(): LLMClientConfiguration? {
+        return AppSettings2.instance.getActiveLLMClientConfiguration(activeLlmClientId)
+    }
 
 }